+++ changelog.md
@@ -16,6 +16,7 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - FilterTable bug fixed, now sets initial filter view after columns have been initialized, for the horizontalAlignment.
 - FilterTableCellEditor.Builder.clickCountToStart() added.
+- DefaultFilterTableCellEditor bug fixed, JCheckBox based editor did not paint selection background, added some DefaultTableCellEditor behaviour.
 ### is.codion.framework.domain
 - EntityType bug fixed, define(List) parameter type changed to List<? extends AttributeDefinition.Builder<?, ?>>.
 ### is.codion.framework.db.rmi